Chore : 바텀시트 리뷰탭

+++ data/src/main/java/com/weit/data/model/place/PlaceReviewListDTO.kt
@@ -6,6 +6,6 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class PlaceReviewListDTO(
     @field:Json(name = "content") val reviews: List<PlaceReviewDTO>,
-    @field:Json(name = "averageRating") val averageRating: Int,
+    @field:Json(name = "averageRating") val averageRating: Float,
     @field:Json(name = "hasNext") val hasNext: Boolean,
 )
