+++ domain/src/main/java/com/weit/domain/usecase/place/GetAverageRatingUseCase.kt
@@ -7,6 +7,6 @@ import javax.inject.Inject
 class GetAverageRatingUseCase @Inject constructor(
     private val repository: PlaceReviewRepository,
 ) {
-    suspend operator fun invoke(placeId: String): Result<Int> =
+    suspend operator fun invoke(placeId: String): Result<Float> =
         repository.getAverageRating(PlaceReviewByPlaceIdInfo(placeId, 20))
 }
